+++ ngo/src/main/java/com/ngo/dto/requestDto/ScrapDto.java
@@ -1,4 +1,4 @@
-package com.ngo.dto;
+package com.ngo.dto.requestDto;
 
 import lombok.Getter;
 import lombok.Setter;
