+++ ngo/src/main/java/com/ngo/dto/requestDto/UserLevelDto.java
@@ -1,4 +1,4 @@
-package com.ngo.dto;
+package com.ngo.dto.requestDto;
 
 import lombok.*;
