+++ ngo/src/main/java/com/ngo/dto/responseDto/AttListDto.java
@@ -1,8 +1,5 @@
-package com.ngo.dto;
+package com.ngo.dto.responseDto;
 
-import com.ngo.model.Attendance;
-import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 
 import java.util.List;
