+++ ngo/src/main/java/com/ngo/dto/responseDto/MemoGetDto.java
@@ -1,4 +1,4 @@
-package com.ngo.dto;
+package com.ngo.dto.responseDto;
 
 import com.ngo.model.Memo;
 import lombok.AccessLevel;
