+++ ngo/src/main/java/com/ngo/dto/responseDto/NewsDto.java
@@ -1,9 +1,24 @@
-package com.ngo.dto;
+package com.ngo.dto.responseDto;
 
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 
+/**
+ * NewsDto is used to construct the response format of Api.
+ * <pre>{@code
+ *     private final String title;
+ *     private final String contents;
+ *     private final String editor;
+ *     private final String media;
+ *     private final String summary;
+ *     private final String thumbnail;
+ * }</pre>
+ * @package : com.ngo.dto
+ * @name : NewsDto.java
+ * @date : 2024. 05. 30.
+ * @author : siyunsmacbook
+*/
 @Getter
 public class NewsDto {
     private final String title;
