+++ ngo/src/main/java/com/ngo/dto/responseDto/RankDto.java
@@ -1,4 +1,4 @@
-package com.ngo.dto;
+package com.ngo.dto.responseDto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
