+++ ngo/src/main/java/com/ngo/dto/responseDto/ScoreDto.java
@@ -1,4 +1,4 @@
-package com.ngo.dto;
+package com.ngo.dto.responseDto;
 
 import com.ngo.model.User;
 import lombok.AccessLevel;
