+++ ngo/src/main/java/com/ngo/dto/responseDto/UserDto.java
@@ -1,4 +1,4 @@
-package com.ngo.dto;
+package com.ngo.dto.responseDto;
 
 import com.ngo.model.User;
 import lombok.AccessLevel;
