+++ ngo/src/main/java/com/ngo/dto/responseDto/WordListDto.java
@@ -1,4 +1,4 @@
-package com.ngo.dto;
+package com.ngo.dto.responseDto;
 
 import lombok.Getter;
